- more work on entitysetup.

+++ JoyceCode/engine/world/CreatorRegistry.cs
@@ -10,9 +10,9 @@ public class CreatorRegistry : AModule
     private GenericIdRegistry<ICreator> _reg =
         new(engine.world.components.Creator.CreatorId_HardcodeMax + 1);
 
-    public uint FindCreatorId(ICreator iCreator) => _reg.FindId(iCreator);
+    public ushort FindCreatorId(ICreator iCreator) => (ushort) _reg.FindId(iCreator);
 
-    public ICreator GetCreator(uint creatorId) => _reg.Get(creatorId);
+    public ICreator GetCreator(ushort creatorId) => _reg.Get((uint)creatorId);
     
     public void UnregisterCreator(ICreator iCreator) => _reg.Unregister(iCreator);
